Separated the node-view in smaller sections.
